Widen the receive window, add IPv6, remove CPU pinning; wire the worker pool into a real transfer path
Three targeted protocol/infra changes plus the security and correctness
hardening that had accumulated alongside them, all verified against the
full test suite (147 tests, fuse_proto_tests + fuse_tests).

## Window size: 64 -> 1024 blocks (wire v5)
kMaxWindow was capped at 255 because window_size was a single wire byte in
SETUP. At 50ms RTT a 64-block window ceilings throughput at ~40 MB/s,
useless on any real WAN path. window_size is now a two-byte field
(setup.cpp put_u16/get_u16, StreamConfig), raising the cap to 1024
in-flight blocks per stream. The receive window is already tracked as a
multi-word bitmask (kMaskWords x uint64), so this only required widening
the wire field and ReceiverStream::window_size_ to uint16_t.

## IPv6 dual-stack support
UdpSocket/PeerAddr were hard-coded to AF_INET/sockaddr_in. PeerAddr now
holds a sockaddr_storage, and open()/resolve() try an address as IPv4
first, then IPv6 ("::" is the v6 wildcard, mirroring "0.0.0.0" for v4).
Added peer_port()/peer_to_string() so callers that need to report an
address don't have to know the family. recv_batch's per-call scratch
(mmsghdr/iovec/sockaddr_storage x 64, ~10KB) moved from the stack to a
socket-owned heap allocation, both to accommodate the larger
sockaddr_storage and to stop paying that stack cost on every hot-path call.

## CPU pinning removed entirely
Benchmarking showed pinning was consistently ~6% slower with no upside
(bench/RESULTS.md), so it's no longer an option rather than a
discouraged one:
- Deleted include/fuse/proto/topology.hpp and src/proto/topology.cpp
  (pin_current_thread_to_core, the NUMA hook, and their test) — nothing
  in production code used them once OrchestratorConfig::pin_to_core,
  least_loaded_core(), and core_load_ bookkeeping were dropped from
  WorkerOrchestrator.
- Dropped the NUMA find_path/find_library block from CMakeLists.txt.
- bench/fuse_bench.cpp no longer runs a "(c) pinned" comparison arm.
- README's benchmark section updated to match.

## Worker pool now drives a real transfer path (fuse::mux_transfer)
Previously WorkerOrchestrator only ran its own synthetic-load tests; the
real send/receive loop lived entirely in transfer.cpp, disconnected from
it. New include/fuse/mux_transfer.hpp + src/proto/mux_transfer.cpp
multiplex several logical streams over one socket with a dynamically
scaled worker pool (send_multiplexed/receive_multiplexed), an alternative
to transfer.hpp's fixed one-thread-per-lane model for callers who want
more streams than threads.

## Security/correctness hardening bundled in (transfer.cpp, sdk.cpp)
- AuthFailed is now actually reachable: classification runs on every exit
  path (idle-timeout, natural completion), not just the happy path, so a
  wrong-key/MITM'd receiver that never advances past auth failures
  correctly reports AuthFailed instead of Timeout.
- Block `flags` moved into the AEAD associated data, so LastBlock can't be
  forged onto an otherwise-authentic encrypted block to truncate a
  transfer early.
- Peer re-anchoring (NAT/CGNAT migration) and RTT sampling now gate on
  ReceiveResult::Accepted (authentic AND new), closing a replay hole where
  a captured datagram resent from a different address could hijack where
  ACKs go.
- HelloAck now carries a server-chosen salt (wire v4) mixed into the
  session key derivation, so a retried/replayed HELLO can no longer make
  two connections derive the same key.
- Integer-overflow-safe bounds checks on wire-supplied offset/length pairs
  (both transfer.cpp and sdk.cpp) — checked as `offset <= size && len <=
  size - offset` instead of `offset + len <= size`, which a huge offset
  could wrap past.
- StreamStart's claimed total_bytes is no longer trusted for allocation
  without a try/catch: a hostile size near UINT64_MAX now fails the lane
  with ResourceLimit instead of taking down the process.
- worst() now ranks lane statuses by severity instead of "last one wins",
  and receive_buffer() no longer hands back a partially-zero-filled buffer
  when any lane failed.
- Sender RTO/timeout tracking split "a local send attempt succeeded" from
  "the peer actually confirmed progress", fixing a case where a dead/
  wrong-keyed peer could make send_lane spin forever instead of honoring
  timeout_ms.
- sdk.cpp: bounded inbox backpressure (256 MiB cap) so a fast sender and a
  slow-reading application can't grow an unbounded queue; an in_use guard
  so fuse_close can't free a connection out from under a call already in
  flight.
- aux.hpp/aux.cpp renamed to control.hpp/control.cpp (AUX is a reserved
  device name on Windows, breaking native Windows checkouts).

Verified: fresh CMake reconfigure + full rebuild, fuse_proto_tests (133/133)
and fuse_tests (14/14) both pass, including new Udp.Ipv6LoopbackRoundTrip
and the MuxTransfer/Sdk suites.

+++ include/fuse/proto/congestion.hpp
@@ -27,7 +27,7 @@ public:
     // `max_window` is the stream's SETUP-negotiated window (the cap).
     // `enabled` mirrors the LOSSLESS flag: a non-lossless stream passes
     // false and its window stays pinned at max_window forever.
-    explicit CongestionController(uint8_t max_window, bool enabled = true,
+    explicit CongestionController(uint16_t max_window, bool enabled = true,
                                   uint32_t clean_windows_to_grow = 3,
                                   uint32_t grow_step = 2)
         : enabled_(enabled),
